requirements.txt
django~=5.1.6
djangorestframework-simplejwt~=5.4.0
djangorestframework~=3.15.2